Enforced manual publishing/subscribing to DB by running:
remove autopublish

Then split up the database files so the main index.js just has to import whatever collections we care to deal with.
Put our publishing function in a separate file as well. Only publish the party if their public or if the user owns the party entry.

Apparently only need to add subscriptions to files where we're retrieving documents from the DB.
Writing to the DB doesn't require the subscribe method.
Now users can see the private parties they've created AND the parties that are public.